Fix injected script wait for hidden state selector
This commit fixes the condition to wait for a 'hidden' state for a JS
element. Now, when the condition is met (the element is no longer
visible), waitForSelector function returns the element, as this is
still a valid element in the DOM, instead of returning undefined.

This will make the call made from the wrapper function
'waitForPredicateFunction':
    return predicateFn(...args) || continuePolling;
evaluate to true and return the JS element instead of hanging on an
infinite loop.