Fix wait for detached selector
This commit fixes waiting for a selector to be in 'detached' state.
Fixes the injected_script wait for selector condition for 'detached'
state, that was returning 'undefined' which eventually caused an
infinite loop from the calling polling function
'waitForPredicateFunction' due to the call:
    return predicateFn(...args) || continuePolling;
As 'undefined' returned from the predicate function resulted in
returning 'continuePolling' from this condition.

Also fixes the frame waitForSelector function implementation in order to
support the returned ElementHandle for the wait to be nil. This is a
valid case only when the state to wait for is 'detached', as in this
case the element is no longer in the DOM, therefore there is no possible
valid ElementHandle to associate to it.